no tricks when proving associativity of min
One of the exercises in chapter 2.4 is proving associativity of min. The text above the exercise suggests that there exists a shorter proof by using one of the newly introduced tricks from this chapter, like the "repeat" tactic for example. However, the offered solution does not feature any of these tricks. An additional text was added to not mislead readers into thinking that any of these tricks can be used here.